requirements.txt
fastapi==0.116.1
gradio==5.49.1
httpx==0.28.1
pydantic==2.11.7
python-dotenv==1.1.1
openai==1.107.3
pytest==8.4.1
matplotlib==3.10.6
